+++ app/src/main/java/com/brohkahn/nasawallpaper/FeedItem.java
@@ -3,13 +3,13 @@ package com.brohkahn.nasawallpaper;
 import java.util.Date;
 
 public class FeedItem {
-    public long id;
+    public int id;
     public String title;
     public String link;
     public String imageName;
     public Date published;
 
-    FeedItem(long id, String title, String link, String imageName, Date published) {
+    FeedItem(int id, String title, String link, String imageName, Date published) {
         this.id = id;
         this.title = title;
         this.link = link;
